revise simd casts

+++ rapidfuzz/details/simd_avx2.hpp
@@ -27,7 +27,7 @@ public:
 
     native_simd(uint64_t a) noexcept
     {
-        xmm = _mm256_set1_epi64x(static_cast<long long int>(a));
+        xmm = _mm256_set1_epi64x(static_cast<int64_t>(a));
     }
 
     native_simd(const uint64_t* p) noexcept
@@ -47,8 +47,8 @@ public:
 
     native_simd load(const uint64_t* p) noexcept
     {
-        xmm = _mm256_set_epi64x(static_cast<long long int>(p[3]), static_cast<long long int>(p[2]),
-                                static_cast<long long int>(p[1]), static_cast<long long int>(p[0]));
+        xmm = _mm256_set_epi64x(static_cast<int64_t>(p[3]), static_cast<int64_t>(p[2]),
+                                static_cast<int64_t>(p[1]), static_cast<int64_t>(p[0]));
         return *this;
     }
 
@@ -116,8 +116,8 @@ public:
 
     native_simd load(const uint64_t* p) noexcept
     {
-        xmm = _mm256_set_epi64x(static_cast<long long int>(p[3]), static_cast<long long int>(p[2]),
-                                static_cast<long long int>(p[1]), static_cast<long long int>(p[0]));
+        xmm = _mm256_set_epi64x(static_cast<int64_t>(p[3]), static_cast<int64_t>(p[2]),
+                                static_cast<int64_t>(p[1]), static_cast<int64_t>(p[0]));
         return *this;
     }
 
@@ -185,8 +185,8 @@ public:
 
     native_simd load(const uint64_t* p) noexcept
     {
-        xmm = _mm256_set_epi64x(static_cast<long long int>(p[3]), static_cast<long long int>(p[2]),
-                                static_cast<long long int>(p[1]), static_cast<long long int>(p[0]));
+        xmm = _mm256_set_epi64x(static_cast<int64_t>(p[3]), static_cast<int64_t>(p[2]),
+                                static_cast<int64_t>(p[1]), static_cast<int64_t>(p[0]));
         return *this;
     }
 
@@ -254,8 +254,8 @@ public:
 
     native_simd load(const uint64_t* p) noexcept
     {
-        xmm = _mm256_set_epi64x(static_cast<long long int>(p[3]), static_cast<long long int>(p[2]),
-                                static_cast<long long int>(p[1]), static_cast<long long int>(p[0]));
+        xmm = _mm256_set_epi64x(static_cast<int64_t>(p[3]), static_cast<int64_t>(p[2]),
+                                static_cast<int64_t>(p[1]), static_cast<int64_t>(p[0]));
         return *this;
     }
